added new table sturcture to elements
